fix autocompaction control test flakiness
The test tries to submit a compaction job after playing
with autocompaction control table switch. However, there is
no reliable way to hook pending compaction task. The code
assumed that with_scheduling_group() closure will never
preempt execution of the stats check.

Sometimes the execution resulted into the following error:

    test/boost/sstable_datafile_test.cc(1076): fatal error: in "autocompaction_control_test":
    critical check cm->get_stats().pending_tasks == 1 || cm->get_stats().active_tasks == 1 has failed

This variant of a fix adds an sstable to the cf, starts
the compaction and awaits until it is finished.

Signed-off-by: Ivan Prisyazhnyy <ivan@scylladb.com>